move types to `eolearn.core.types`
