+++ core/eolearn/core/eodata_merge.py
@@ -25,8 +25,8 @@ from sentinelhub import BBox
 
 from .constants import FeatureType
 from .exceptions import EORuntimeWarning
+from .types import Literal
 from .utils.parsing import FeatureParser, FeatureSpec, FeaturesSpecification
-from .utils.types import Literal
 
 if TYPE_CHECKING:
     from .eodata import EOPatch
