+++ core/eolearn/core/eotask.py
@@ -23,6 +23,7 @@ from dataclasses import dataclass
 from typing import Any, Dict, Iterable, Type, TypeVar, Union
 
 from .constants import FeatureType
+from .types import EllipsisType
 from .utils.parsing import (
     FeatureParser,
     FeaturesSpecification,
@@ -31,7 +32,6 @@ from .utils.parsing import (
     parse_renamed_feature,
     parse_renamed_features,
 )
-from .utils.types import EllipsisType
 
 LOGGER = logging.getLogger(__name__)
