+++ core/eolearn/tests/test_utils/test_parsing.py
@@ -8,8 +8,8 @@ import pytest
 from sentinelhub import CRS, BBox
 
 from eolearn.core import EOPatch, FeatureParser, FeatureType
+from eolearn.core.types import EllipsisType
 from eolearn.core.utils.parsing import FeatureRenameSpec, FeatureSpec, FeaturesSpecification
-from eolearn.core.utils.types import EllipsisType
 
 
 @dataclass
