+++ io/eolearn/io/sentinelhub_process.py
@@ -36,8 +36,8 @@ from sentinelhub import (
 from sentinelhub.type_utils import RawTimeIntervalType
 
 from eolearn.core import EOPatch, EOTask, FeatureType, FeatureTypeSet
+from eolearn.core.types import Literal
 from eolearn.core.utils.parsing import FeatureRenameSpec, FeatureSpec, FeaturesSpecification
-from eolearn.core.utils.types import Literal
 
 LOGGER = logging.getLogger(__name__)
